+++ packages/apollo-server-core/src/__tests__/runHttpQuery.test.ts
@@ -58,4 +58,95 @@ describe('runHttpQuery', () => {
       });
     });
   });
+
+  describe('when allowBatchedHttpRequests is false', () => {
+    const mockDisabledBatchQueryRequest = {
+      method: 'GET',
+      query: {
+        query: '{ testString }',
+      },
+      options: {
+        debug: false,
+        schema,
+        schemaHash: 'deprecated' as SchemaHash,
+        allowBatchedHttpRequests: false,
+      },
+      request: new MockReq(),
+    };
+
+    it('succeeds when there are no batched queries in the request', async () => {
+      await expect(
+        runHttpQuery([], mockDisabledBatchQueryRequest),
+      ).resolves.not.toThrow();
+    });
+
+    it('throws when there are batched queries in the request', () => {
+      const batchedQueryRequest = Object.assign(
+        {},
+        mockDisabledBatchQueryRequest,
+        {
+          query: [
+            {
+              query: '{ testString }',
+            },
+            {
+              query: '{ testString }',
+            },
+          ],
+        },
+      );
+      return runHttpQuery([], batchedQueryRequest).catch(
+        (err: HttpQueryError) => {
+          expect(err.statusCode).toEqual(400);
+          expect(err.message).toEqual(
+            JSON.stringify({
+              errors: [
+                {
+                  message: 'Operation batching disabled.',
+                  extensions: { code: 'INTERNAL_SERVER_ERROR' },
+                },
+              ],
+            }) + '\n',
+          );
+        },
+      );
+    });
+  });
+
+  describe('when allowBatchedHttpRequests is true', () => {
+    const mockEnabledBatchQueryRequest = {
+      method: 'GET',
+      query: {
+        query: '{ testString }',
+      },
+      options: {
+        debug: false,
+        schema,
+        schemaHash: 'deprecated' as SchemaHash,
+        allowBatchedHttpRequests: true,
+      },
+      request: new MockReq(),
+    };
+
+    it('succeeds when there are multiple queries in the request', async () => {
+      const multipleQueryRequest = Object.assign(
+        {},
+        mockEnabledBatchQueryRequest,
+        {
+          query: [
+            {
+              query: '{ testString }',
+            },
+            {
+              query: '{ testString }',
+            },
+          ],
+        },
+      );
+
+      await expect(
+        runHttpQuery([], multipleQueryRequest),
+      ).resolves.not.toThrow();
+    });
+  });
 });