+++ packages/apollo-server-integration-testsuite/src/index.ts
@@ -878,6 +878,43 @@ export default ({
         });
       }, 3000); // this test will fail due to timeout if running serially.
 
+      it('disables batch requests when allowBatchedHttpRequests is false', async () => {
+        app = await createApp({
+          graphqlOptions: {
+            schema,
+            allowBatchedHttpRequests: false,
+          },
+        });
+
+        const res = await request(app)
+          .post('/graphql')
+          .send([
+            {
+              query: `
+                      query test($echo: String){ testArgument(echo: $echo) }
+                      query test2{ testString }`,
+              variables: { echo: 'world' },
+              operationName: 'test2',
+            },
+            {
+              query: `
+                      query testX($echo: String){ testArgument(echo: $echo) }`,
+              variables: { echo: 'yellow' },
+              operationName: 'testX',
+            },
+          ]);
+
+        expect(res.status).toEqual(400);
+        expect(res.body).toEqual({
+          errors: [
+            {
+              message: 'Operation batching disabled.',
+              extensions: { code: 'INTERNAL_SERVER_ERROR' },
+            },
+          ],
+        });
+      });
+
       it('clones batch context', async () => {
         app = await createApp({
           graphqlOptions: {